+++ R/QR_functions.R
@@ -252,7 +252,7 @@ compute_score.QR_matrix <- function(
     na.rm = TRUE,
     n_contrib_score = NULL,
     conditional_indicator = NULL,
-    thresholds = getOption("jdc_thresholds"),
+    thresholds = getOption("JDCruncheR.thresholds"),
     ...
 ) {
     if (!all(names(score_pond) %in% colnames(x[["modalities"]]))) {