+++ R/zzz.R
@@ -1,6 +1,6 @@
 #' @keywords internal
 .onLoad <- function(libname, pkgname) {
-    if (is.null(getOption("jdc_thresholds"))) {
+    if (is.null(getOption("JDCruncheR.thresholds"))) {
         set_thresholds()
     }
 }